requirements.txt
feedparser==6.0.10
requests==2.31.0
python-docx==0.8.11
python-dateutil==2.8.2
beautifulsoup4==4.12.2
lxml==4.9.3
